+++ modules/apps/portal-configuration/portal-configuration-module-configuration/src/main/java/com/liferay/portal/configuration/module/configuration/internal/configuration/PortletPreferencesConfigurationListener.java
@@ -1,0 +1,37 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2023 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.portal.configuration.module.configuration.internal.configuration;
+
+import com.liferay.petra.string.StringPool;
+import com.liferay.portal.configuration.module.configuration.internal.ConfigurationOverrideInstance;
+import com.liferay.portal.kernel.util.StringUtil;
+
+import org.osgi.service.cm.ConfigurationEvent;
+import org.osgi.service.cm.ConfigurationListener;
+import org.osgi.service.component.annotations.Component;
+
+/**
+ * @author Joao Victor Alves
+ */
+@Component(service = ConfigurationListener.class)
+public class PortletPreferencesConfigurationListener
+	implements ConfigurationListener {
+
+	@Override
+	public void configurationEvent(ConfigurationEvent configurationEvent) {
+		String key = configurationEvent.getPid();
+
+		String factoryPid = configurationEvent.getFactoryPid();
+
+		if (factoryPid != null) {
+			key = StringUtil.replaceLast(
+				factoryPid, ".scoped", StringPool.BLANK);
+		}
+
+		ConfigurationOverrideInstance.clearConfigurationOverrideInstance(key);
+	}
+
+}
